--- client/internal/tui/simple_input.go
@@ -1,14 +1,0 @@
-package tui
-
-type textinput struct {
-	placeholder string
-	value       string
-}
-
-func (t *textinput) Value() string {
-	return t.value
-}
-
-func (t *textinput) SetValue(s string) {
-	t.value = s
-}
